crud clientes

+++ TCC.Domain/Entities/Cliente.cs
@@ -4,7 +4,7 @@ namespace TCC.Domain.Entities
 {
     public class Cliente : EntityBase
     {
-        public int Cpf { get; set; }
+        public long Cpf { get; set; }
         public string Nome { get; set; }
     }
 }
